Support for Linux cooked-mode capture (SLL)
Implement support for Linux cooked-mode capture (SLL). Add a command
line option to specify that the source pcap file contains SLL header.

See http://wiki.wireshark.org/SLL

This is the pseudo-protocol used by libpcap on Linux to capture from the
"any" device and to capture on some devices where the native link layer
header isn't available or can't be used.

When capturing from the "any" device, or from one of those other
devices, in Linux, the libpcap doesn't supply the link-layer header for
the real "hardware protocol" like Ethernet, but instead supplies a fake
link-layer header for this pseudo-protocol.